app/requirements.txt
anyio
click
ecdsa
fastapi
h11
idna
mangum
numpy
pandas
pyasn1
pydantic
pytz
rsa
six
sniffio
starlette
typing_extensions
tzdata
uvicorn
python-jose
slowapi
python-multipart
requests
boto3